pyproject.toml
[tool.poetry]
name = "omegaproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "23.3.0"
click = "8.1.3"
colorama = "0.4.6"
flask = "2.2.2"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.2"
mypy-extensions = "1.0.0"
mysql-connector-python = "8.0.32"
packaging = "23.0"
pathspec = "0.11.1"
platformdirs = "3.2.0"
protobuf = "3.20.3"
werkzeug = "2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
